Listの中の単一のメモを表示する部分をMemoとして切り出した
・構造的にわかりやすいため単一のメモの部分を切り出した。それに伴って、既存のMemoコンポーネントの名称をMemosSectionへ変更した。
