+++ src/List.jsx
@@ -1,19 +1,18 @@
 import React from 'react'
 import PropTypes from 'prop-types'
+import Memo from './Memo.jsx'
 
 export default class List extends React.Component {
   render () {
     return (
       <ul>
         {this.props.memos.map((memo, index) => (
-          <li key={index}>
-            <button
-              className="existingMemo"
-              onClick={(event) => this.props.displayForm(event, index)}
-            >
-              {memo[0]}
-            </button>
-          </li>
+          <Memo
+            key={index}
+            index={index}
+            displayForm={this.props.displayForm}
+            memo={memo}
+          />
         ))}
       </ul>
     )
